+++ src/Console/Facades/Console.php
@@ -10,6 +10,7 @@ use Simtabi\Laranail\Console\Prompter\Prompter;
 use Simtabi\Laranail\Console\Tools\Formatting\ConsoleUIFormatter;
 use Simtabi\Laranail\Console\Tools\Support\Capabilities;
 use Simtabi\Laranail\Console\Tools\Support\Color;
+use Simtabi\Laranail\Console\Tools\Support\Emoji;
 use Simtabi\Laranail\Console\Tools\Widgets\Banner;
 use Simtabi\Laranail\Console\Tools\Widgets\Box;
 use Simtabi\Laranail\Console\Tools\Widgets\Gauge;
@@ -39,6 +40,7 @@ use Symfony\Component\Console\Output\OutputInterface;
  * @method static StepFlow steps(array $steps = [])
  * @method static TaskProgress tasks(?OutputInterface $output = null)
  * @method static Color color()
+ * @method static Emoji emoji()
  * @method static Capabilities capabilities()
  * @method static Prompter prompter()
  *